Add RC4-SHA1 and AES-GCM cipher suites
Abstract some of the details of various cipher suites away from the core
code.

Added AES-GCM since it's the most efficient 'seekable' cipher suite
suitable for DTLS. RC4 won't cut it since datagrams can get re-ordered
in transit so the recipient won't know where in the stream to decode
from.

AES-GCM also weighs in as the lightest 'high security' suite out there
which will be safe in the soon to be coming TLS-1.3 (where only AEAD
ciphers are allowed) and the currently pending RC4-is-forbidden world.

Since we have the machinery now to support more than one cipher suite, we
may as well add in RC4-SHA1, since we already have SHA1. Unfortunately
the problem with RC4-MD5 is that messages can be decrypted without
having to break the MAC check. So adding a stronger MAC doesn't give us
much. But it pretty much comes for free, so why not?